fix: Decimal serialization error
The change from Float to Numeric in CitiBuy model tables caused a 
serialization error to be thrown by 
ContractManagement.update_contract_list(). This commit fixes that issue 
by reverting them back to Float.

+++ app/src/dgs_fiscal/systems/citibuy/models/invoice_tables.py
@@ -14,7 +14,7 @@ class Invoice(db.Base):
     invoice_nbr = db.Column("INVOICE_NBR", db.String)
     invoice_date = db.Column("INVOICE_DATE", db.DateTime)
     status = db.Column("INVOICE_STATUS", db.String)
-    amount = db.Column("INVOICE_AMT", db.Numeric(precision=2))
+    amount = db.Column("INVOICE_AMT", db.Float(precision=2))
     vendor_id = db.Column(
         "VENDOR_NBR",
         db.String,
